refactor: HeroArea 의 validateTime을 유틸로 분리
